+++ DawnLib.Dusk/src/DuskKeys.cs
@@ -1,0 +1,11 @@
+using Dawn;
+
+namespace Dusk;
+
+public static partial class DuskKeys
+{
+    public const string Namespace = "dawn_lib";
+
+    // Data Keys
+    public static readonly NamespacedKey EntityReplacements = NamespacedKey.From(Namespace, "entity_replacements");
+}