+++ DawnLib/src/API/TerminalCommands/.TerminalCommandRegistration.cs
@@ -237,10 +237,10 @@ static class TerminalCommandRegistration
 
     private static string ReplaceInternalLevelNames(string input)
     {
-        string result = string.Empty;
+        string result = input;
         foreach ((string internalName, string humanName) in _internalToHumanRouteNames)
         {
-            result = input.Replace(internalName, humanName);
+            result = result.Replace(internalName, humanName);
         }
         return result;
     }